requirements.txt
requests~=2.32.5
python-dotenv~=1.2.1
pandas~=2.3.3
numpy~=2.3.5

# MySQL / PGSql
SQLAlchemy~=2.0.45
psycopg~=3.3.2
# Mongo
pymongo~=4.15.5
# Binance API
python-binance~=1.0.33
# API
fastapi~=0.115.0
uvicorn[standard]~=0.32.0
# WebSocket
websocket-client~=1.8.0
# Dashboard
dash~=4.0.0
dash-bootstrap-components~=2.0.4
plotly~=6.5.2
# Scheduler (sera remplacé par Airflow)
schedule~=1.2.2
